ai_pipeline/requirements.txt
ollama
pytesseract
Pillow
requests
tqdm
langchain
langchain-community
chromadb
easyocr
torch
torchvision
transformers
pdf2image
pymupdf
pdfplumber
